requirements.txt
absl-py==0.15.0
addict==2.4.0
aim==3.5.4
aimrecords==0.0.7
aimrocks==0.0.10
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
apex==0.1
asgiref==3.5.0
astor==0.8.1
astunparse==1.6.3
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.1
attrs==21.4.0
av==10.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
base58==2.0.1
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
captum==0.4.1
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
cloudpickle==2.0.0
coconut==1.5.0
colorama==0.4.4
coverage==5.5
cox==0.1.post3
cryptography==3.4.8
cxxfilt==0.3.0
cycler==0.11.0
Cython==0.29.25
cytoolz==0.11.0
dask==2021.10.0
decorator==5.1.1
dicttoxml==1.7.4
dill==0.3.4
easydict==1.7
einops==0.4.0
et-xmlfile==1.1.0
fastapi==0.67.0
filelock==3.6.0
fire==0.4.0
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.42.0
h11==0.13.0
h5py==2.10.0
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
importlib-metadata==6.8.0
importlib-resources==5.4.0
imutils==0.5.4
iniconfig==1.1.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json-tricks==3.15.5
jsonpatch==1.32
jsonpointer==2.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.5.8
lime==0.2.0.1
locket==0.2.1
loguru==0.5.3
Mako==1.1.6
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.0
matplotlib==3.5.0
matplotlib-inline==0.1.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv==2.0.1
mmengine==0.8.2
mock==4.0.3
monitor-model==0.0.0
more-itertools==8.8.0
multidict==5.2.0
munkres==1.1.4
networkx==2.6.3
nibabel==3.2.1
ninja==1.10.2.2
nltk==3.6.5
numexpr==2.8.1
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.2.6
onnxruntime==1.8.0
opencv-python==4.7.0.68
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
PIMS==0.6.1
platformdirs==3.8.1
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.21
pydantic==1.9.0
pyDeprecate==0.3.2
pydicom==2.2.2
Pygments==2.15.1
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyquaternion==0.9.9
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2022.7.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.2
requests==2.27.1
requests-oauthlib==1.3.0
RestrictedPython==5.2
retry==0.9.2
rich==13.4.2
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.1
Shapely==1.8.2
sip==4.19.13
six==1.16.0
sklearn-contrib-lightning==0.6.0
slicerator==1.1.0
smmap==4.0.0
SQLAlchemy==1.4.31
starlette==0.14.2
tables==3.6.1
tabulate==0.8.9
tcav==0.2.1
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorboardX==2.0
tensorflow==2.4.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
tf-slim==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.9.1.post3
torchaudio==0.9.0a0+a85b239
torchfile==0.1.0
torchmetrics==0.7.3
torchnet==0.0.4
torchvision==0.10.0a0
tornado==6.1
tqdm==4.43.0
traitlets==5.1.1
ttach==0.0.3
typing_extensions==4.7.1
urllib3==1.26.8
uvicorn==0.17.5
vector-quantize-pytorch==0.1.0
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==2.0.2
wrapt==1.13.3
xjm==0.0.4
xmltodict==0.12.0
yacs==0.1.6
yapf==0.40.1
yarl==1.6.3
zipp==3.7.0
